pyproject.toml
[project]
name = "sensorevalgui"
version = "0.1.0"
description = "GUI tool for image-sensor evaluation"
authors = [{name = "Magik Eye Inc."}]
dependencies = [
    "numpy>=1.25,<2.0",
    "matplotlib>=3.8,<4.0",
    "tifffile>=2023.7",
    "PySide6>=6.6,<6.7",
    "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core", "gui", "utils"]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
